+++ partiql-spi/src/main/kotlin/org/partiql/spi/function/builtins/FnEq.kt
@@ -16,8 +16,8 @@ import org.partiql.spi.value.Datum
  * > Equality never fails in the type-checking mode and never returns MISSING in the permissive mode. Instead, it can
  * compare values of any two types, according to the rules of the PartiQL type system. For example, 5 = 'a' is false.
  *
- * For the existing conformance tests, when an operand is NULL, the output is NULL. When an operand is MISSING,
- * the output is MISSING (missing value propagation). This implementation follows the existing conformance tests.
+ * For the existing conformance tests, whenever an operand is NULL or MISSING, the output is NULL. This implementation
+ * follows this.
  *
  * TODO: The PartiQL Specification needs to clearly define the semantics of MISSING. That being said, this implementation
  *  follows the existing conformance tests and SQL:1999.
@@ -26,14 +26,12 @@ private val name = FunctionUtils.hide("eq")
 internal val FnEq = FnOverload.Builder(name)
     .addParameters(PType.dynamic(), PType.dynamic())
     .returns(PType.bool())
-    .isNullCall(false)
+    .isNullCall(true)
     .isMissingCall(false)
     .body { args ->
         val lhs = args[0]
         val rhs = args[1]
         if (lhs.isMissing || rhs.isMissing) {
-            Datum.missing(PType.bool())
-        } else if (lhs.isNull || rhs.isNull) {
             Datum.nullValue(PType.bool())
         } else {
             Datum.bool(Datum.comparator().compare(lhs, rhs) == 0)